+++ packages/graphql/src/SortBy/Operators/Field.php
@@ -12,11 +12,11 @@ use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\TypeProvider;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\TypeSource;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Exceptions\OperatorUnsupportedBuilder;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Property;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Direction;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Eloquent\Builder as EloquentSorter;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Query\Builder as QuerySorter;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Scout\Builder as ScoutSorter;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Sorter;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\Direction;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\EloquentSorter;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\QuerySorter;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\ScoutSorter;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\Sorter;
 use LastDragon_ru\LaraASP\GraphQL\SortBy\Types\Direction as DirectionType;
 use Nuwave\Lighthouse\Execution\Arguments\Argument;
 use Override;
