+++ packages/graphql/src/SortBy/Operators/FieldTest.php
@@ -8,11 +8,11 @@ use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use Laravel\Scout\Builder as ScoutBuilder;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Property;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Direction;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Eloquent\Builder as EloquentSorter;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Query\Builder as QuerySorter;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Scout\Builder as ScoutSorter;
 use LastDragon_ru\LaraASP\GraphQL\SortBy\Directives\Directive;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\Direction;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\EloquentSorter;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\QuerySorter;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\ScoutSorter;
 use LastDragon_ru\LaraASP\GraphQL\Testing\Package\DataProviders\BuilderDataProvider;
 use LastDragon_ru\LaraASP\GraphQL\Testing\Package\TestCase;
 use LastDragon_ru\LaraASP\Testing\Providers\ArrayDataProvider;
