+++ packages/graphql/src/SortBy/Sorters/ScoutSorter.php
@@ -1,18 +1,16 @@
 <?php declare(strict_types = 1);
 
-namespace LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Scout;
+namespace LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters;
 
 use Laravel\Scout\Builder as ScoutBuilder;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\Scout\FieldResolver;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Property;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Direction;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Builders\Sorter;
 use Override;
 
 /**
  * @implements Sorter<ScoutBuilder>
  */
-class Builder implements Sorter {
+class ScoutSorter implements Sorter {
     public function __construct(
         protected FieldResolver $fieldResolver,
     ) {
